#14841 Move extension actions to browser (cont..)
